requirements.txt
filelock~=3.6.0
pandas~=1.4.1
numpy~=1.19.5
tensorflow~=2.6.3
nltk~=3.7
tqdm~=4.63.0
sib-clustering~=0.1.6
scikit-learn~=1.0.2
transformers~=4.17.0
matplotlib~=3.5.1
